backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
python-multipart==0.0.6
weasyprint==62.3
# pydyf==0.11.0
markdown==3.5.1
pyyaml==6.0.1
langchain-openai==0.2.14
langchain-anthropic==0.3.8
